Some more cleanups.
Jeanphi

+++ cameras/environment.cpp
@@ -112,11 +112,11 @@ bool EnvironmentCamera::GetSamplePosition(const Point &p, const Vector &wi, floa
 		return false;
 	const Vector w = WorldToCamera(wi);
 	const float cosTheta = w.y;
-	const float theta = acos(min(1.f, cosTheta));
+	const float theta = acosf(min(1.f, cosTheta));
 	*y = theta * film->yResolution * INV_PI;
 	const float sinTheta = sqrtf(Clamp(1.f - cosTheta * cosTheta, 1e-5f, 1.f));
 	const float cosPhi = w.x / sinTheta;
-	const float phi = acos(Clamp(cosPhi, -1.f, 1.f));
+	const float phi = acosf(Clamp(cosPhi, -1.f, 1.f));
 	if (w.z >= 0.f)
 		*x = phi * film->xResolution * INV_TWOPI;
 	else
